initial commit
Contains the ruby script and json file containing key data. I originally had the script in a bin folder in my home, with the json in a src folder within the bin, for easy access. If you would like to change the location simply edit the ruby file and change the json location.